+++ R/reexport.R
@@ -1,0 +1,5 @@
+# REEXPORT FROM ARKHE
+
+#' @importFrom arkhe bootstrap
+#' @export
+arkhe::bootstrap